backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn>=0.23.2
aiofiles>=23.2.1
jinja2>=3.1.2
python-multipart>=0.0.6
httpx>=0.25.1
aiohttp>=3.8.5
gunicorn==20.1.0
websockets==12.0

# Data handling and serialization
pydantic>=2.4.2
pydantic-settings==2.1.0
sqlalchemy>=2.0.23
sqlalchemy-utils>=0.41.1
python-dotenv>=1.0.0
requests>=2.31.0
jose>=1.0.0
pyjwt>=2.8.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# AI and LLM
openai>=1.1.1
langchain>=0.0.335
tiktoken>=0.5.1
regex>=2023.10.3
nltk>=3.8.1

# Data science
pandas==2.2.0
numpy==1.26.4

# News and feed parsing
feedparser>=6.0.10
beautifulsoup4>=4.12.2
tqdm>=4.66.1
markdown>=3.5.1

# External APIs
ccxt==4.2.15
python-binance==1.0.19
twilio==8.12.0
python-telegram-bot==20.8
praw>=7.7.1

# Cache and storage
redis>=5.0.1
python-dateutil==2.9.0.post0

# Testing and dev tools
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0

# System utilities
psutil>=5.9.6